requirements.txt
# Core Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# ML & NLP
sentence-transformers>=2.3.1
transformers>=4.37.2
torch>=2.0.0
numpy>=1.24.0

# Vector Search
faiss-cpu>=1.7.4

# Data Processing
scikit-learn>=1.3.0
pandas>=2.0.0

# API & Utilities
python-multipart==0.0.6
httpx==0.26.0
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Logging & Monitoring
loguru==0.7.2
